backend/requirements.txt
Flask>=2.2
flask-cors>=4.0
openai>=1.0
requests>=2.31
python-dotenv>=1.0
